% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_personfolder.R
\name{check_personfolder}
\alias{check_personfolder}
\title{check_personfolder}
\usage{
check_personfolder(personfolder)
}
\arguments{
\item{personfolder}{personfolder}
}
\value{
No output
}
\description{
check_personfolder
}
