% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAppActive.R
\name{getAppActive}
\alias{getAppActive}
\title{getAppActive}
\usage{
getAppActive(filename)
}
\arguments{
\item{filename}{name of file where pdk-foreground-application is stored (txt).}
}
\value{
timestamps (POSIX) on which the apps were on.
}
\description{
getAppActive
}
