% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMovementGPS.R
\name{getMovementGPS}
\alias{getMovementGPS}
\title{getMovementGPS}
\usage{
getMovementGPS(filename, desiredtz)
}
\arguments{
\item{filename}{name of file where pdk-location is stored (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
timestamps (POSIX) on which the phone moved based on GPS/Speed information.
}
\description{
getMovementGPS
}
