% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMovementPSG.R
\name{getMovementPSG}
\alias{getMovementPSG}
\title{getAppActive}
\usage{
getMovementPSG(filename)
}
\arguments{
\item{filename}{name of file where pdk-location is stored (txt).}
}
\value{
timestamps (POSIX) on which the phone moved based on PSG/Speed information.
}
\description{
getAppActive
}
