% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreenState.R
\name{getScreenState}
\alias{getScreenState}
\title{getScreenState}
\usage{
getScreenState(filename)
}
\arguments{
\item{filename}{name of file where pdk-screen-state is stored (txt).}
}
\value{
timestamps (POSIX) on which the screen was on.
}
\description{
getScreenState
}
