% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSurvey.R
\name{getSurvey}
\alias{getSurvey}
\title{getSurvey}
\usage{
getSurvey(filename, desiredtz)
}
\arguments{
\item{filename}{name of file where pdk-system-status is stored (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
TimeStamps and Answers survey.
}
\description{
getSurvey
}
