% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemHalted.R
\name{getSystemHalted}
\alias{getSystemHalted}
\title{getSystemHalted}
\usage{
getSystemHalted(filename, desiredtz)
}
\arguments{
\item{filename}{name of file where pdk-system-status is stored (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
TimeStamps when phone was halted.
}
\description{
getSystemHalted
}
