% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWithingsActivity.R
\name{getWithingsActivity}
\alias{getWithingsActivity}
\title{getWithingsActivity}
\usage{
getWithingsActivity(filefolder, desiredtz)
}
\arguments{
\item{filefolder}{path to folder Withings-.... with Withings data files (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
data.frame with timestamps (POSIX) on which body info was entered and/or there was movement.
}
\description{
getWithingsActivity
}
