% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heatmaps.R
\name{heatmaps}
\alias{heatmaps}
\title{heatmaps}
\usage{
heatmaps(Dshort, Dlong, heatmapsfile, heatmapsfile_steps,
  simplify.behavioralclasses, Dsurvey = c(), startDate = c(),
  endDate = c(), desiredtz)
}
\arguments{
\item{Dshort}{data.frame generated by agg.sleepsight}

\item{Dlong}{data.frame generated by agg.sleepsight}

\item{heatmapsfile}{png.file to which the plot should be written to}

\item{heatmapsfile_steps}{png.file to which the plot should be written to}

\item{simplify.behavioralclasses}{TRUE or FALSE}

\item{Dsurvey}{data.frame generated by agg.sleepsight}

\item{startDate}{Character,  optional argument to specify the startDate of the recording, e.g. 2019-01-01}

\item{endDate}{Character, optional argument to specify the startDate of the recording, e.g. 2019-01-01}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
no output, just a file is stored
}
\description{
heatmaps
}
