% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iso8691chartime2POSIX.R
\name{iso8601chartime2POSIX}
\alias{iso8601chartime2POSIX}
\title{iso8601chartime2POSIX}
\usage{
iso8601chartime2POSIX(x, tz)
}
\arguments{
\item{x}{timestamp (character) in iso801 format}

\item{tz}{timezone (character) in Europe/London format}
}
\value{
timestamp in POSIX format
}
\description{
iso8601chartime2POSIX
}
\examples{
iso8601chartime2POSIX(x="2017-05-07T13:00:00+0200",tz="Europe/Amsterdam")
}
