% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plot_timeseries}
\usage{
plot_timeseries(
  D24HR,
  Dsurvey,
  timeseriesfile,
  desiredtz,
  startDate = c(),
  endDate = c()
)
}
\arguments{
\item{D24HR}{data.frame with corresponding name produced by agg.sleepsight.}

\item{Dsurvey}{data.frame with corresponding name produced by agg.sleepsight.}

\item{timeseriesfile}{png.file to which the plot should be written to}

\item{desiredtz}{timezone (character) in Europe/London format}

\item{startDate}{Character,  optional argument to specify the startDate of the recording, e.g. 2019-01-01}

\item{endDate}{Character, optional argument to specify the startDate of the recording, e.g. 2019-01-01}
}
\value{
no output, just a file is stored
}
\description{
plot_timeseries
}
