% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replaceVarWithSpace.R
\name{replaceVarWithSpace}
\alias{replaceVarWithSpace}
\title{replaceVarWithSpace}
\usage{
replaceVarWithSpace(x)
}
\arguments{
\item{x}{data.frame}
}
\value{
same data.frame but with specific column names with a space in them replaced by column names with a dot
}
\description{
replaceVarWithSpace
}
