% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise.R
\name{summarise}
\alias{summarise}
\title{summarise}
\usage{
summarise(outputfolder, csvfile, desiredtz)
}
\arguments{
\item{outputfolder}{RData datafile where the output of preprocess is stored.}

\item{csvfile}{file name where data overview needs to be stored}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
List with two data frames: D24HR and D5min
}
\description{
summarise
}
