% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testplot.R
\name{testplot}
\alias{testplot}
\title{testplot}
\usage{
testplot(outputfolder, csvfile)
}
\arguments{
\item{outputfolder}{output folder where to store the plot and where to find the csv file}

\item{csvfile}{csvfile with the aggregated data}
}
\value{
No output, a png file is stored in the outputfolder
}
\description{
testplot
}
