% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addPOSIX.R
\name{addPOSIX}
\alias{addPOSIX}
\title{addPOSIX}
\usage{
addPOSIX(x, desiredtz)
}
\arguments{
\item{x}{data.frame with column "Created.Date" that has timestamps in sleepsight format}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
data.frame x with timestamps added in POSIX format
}
\description{
addPOSIX
}
