% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.sleepsight.R
\name{agg.sleepsight}
\alias{agg.sleepsight}
\title{agg.sleepsight}
\usage{
agg.sleepsight(outputfolder, csvfile, desiredtz, minmisratio = 1/3,
  shortwindow = 1, longwindow = 30)
}
\arguments{
\item{outputfolder}{RData datafile where the output of preprocess is stored.}

\item{csvfile}{file name where data overview needs to be stored}

\item{desiredtz}{timezone (character) in Europe/London format}

\item{minmisratio}{minimum missing ratio per day, day will be ignored if more data is missing}

\item{shortwindow}{short window length in minutes to aggregate to}

\item{longwindow}{long window length in minutes to aggregate to}
}
\value{
List with data frames that hold the aggregated data: D24HR, Dshort, Dlong, and Dsurvey.
}
\description{
agg.sleepsight
}
