% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg.sleepsight.R
\name{agg.sleepsight}
\alias{agg.sleepsight}
\title{agg.sleepsight}
\usage{
agg.sleepsight(aggregatefile, csvfile, surveyfile, desiredtz,
  minmisratio = 1/3, shortwindow = 1, longwindow = 30,
  withings.mode = "dd")
}
\arguments{
\item{aggregatefile}{RData file where the output of aggregation will be stored.}

\item{csvfile}{csv file where data overview was stored}

\item{surveyfile}{RData file where the preprocessed survey data is stored}

\item{desiredtz}{timezone (character) in Europe/London format}

\item{minmisratio}{minimum missing ratio per day, day will be ignored if more data is missing}

\item{shortwindow}{short window length in minutes to aggregate to}

\item{longwindow}{long window length in minutes to aggregate to}

\item{withings.mode}{character ("pdk", "dd") to indicate whether to prioritise pdk or dd}
}
\value{
List with data frames that hold the aggregated data: D24HR, Dshort, Dlong, and Dsurvey.
}
\description{
agg.sleepsight
}
