% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2csv.R
\name{export2csv}
\alias{export2csv}
\title{export2csv}
\usage{
export2csv(outputfolder, csvfile, desiredtz)
}
\arguments{
\item{outputfolder}{RData datafile where the output of preprocess is stored.}

\item{csvfile}{file name where data overview needs to be stored}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
no output, just a file is stored
}
\description{
export2csv
}
