% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBatInteract.R
\name{getBatInteract}
\alias{getBatInteract}
\title{getBatInteract}
\usage{
getBatInteract(filename)
}
\arguments{
\item{filename}{name of file where pdk-device-battery is stored (txt).}
}
\value{
timestamps (POSIX) on which the battery was either being connected or unplugged.
}
\description{
getBatInteract
}
