% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPhoneAcc.R
\name{getPhoneAcc}
\alias{getPhoneAcc}
\title{getPhoneAcc}
\usage{
getPhoneAcc(filefolder, desiredtz)
}
\arguments{
\item{filefolder}{path to folder pdk-sensor-accelerometer with Phone acceleration files (txt).}
}
\value{
acc a dataframe with timestamps (POSIX), acceleration, sf sample frequency and source.
}
\description{
getPhoneAcc
}
