% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getScreenState.R
\name{getScreenState}
\alias{getScreenState}
\title{getScreenState}
\usage{
getScreenState(filename, desiredtz)
}
\arguments{
\item{filename}{name of file where pdk-screen-state is stored (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
timestamps (POSIX) on which the screen was on.
}
\description{
getScreenState
}
