% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSunSetRise.R
\name{getSunSetRise}
\alias{getSunSetRise}
\title{getSunSetRise}
\usage{
getSunSetRise(filename, desiredtz)
}
\arguments{
\item{filename}{name of file where pdk-time-of-day is stored (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
data.frame with timestamps (POSIX) for detected events (sunset/sunrise).
}
\description{
getSunSetRise
}
