% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWithingsData.R
\name{getWithingsData}
\alias{getWithingsData}
\title{getWithingsData}
\usage{
getWithingsData(filefolder, desiredtz)
}
\arguments{
\item{filefolder}{path to folder Withings-.... with Withings data files (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
data.frame with timestamps (POSIX) on which body info was entered and/or there was movement.
}
\description{
getWithingsData
}
