% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWithingsSleep.R
\name{getWithingsSleep}
\alias{getWithingsSleep}
\title{getWithingsSleep}
\usage{
getWithingsSleep(filefolder, desiredtz)
}
\arguments{
\item{filefolder}{path to folder Withings-.... with Withings data files (txt).}

\item{desiredtz}{timezone (character) in Europe/London format}
}
\value{
data.frame with timestamps (POSIX) for detected sleep stages by Withings.
}
\description{
getWithingsSleep
}
