% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{plot_timeseries}
\usage{
plot_timeseries(D24HR, Dsurvey, timeseriesfile)
}
\arguments{
\item{D24HR}{data.frame with corresponding name produced by agg.sleepsight.}

\item{Dsurvey}{data.frame with corresponding name produced by agg.sleepsight.}

\item{timeseriesfile}{png.file to which the plot should be written to}
}
\value{
no output, just a file is stored
}
\description{
plot_timeseries
}
