% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{preprocess}
\usage{
preprocess(
  personfolder,
  desiredtz,
  overwrite = FALSE,
  outputfolder,
  ignore.light = TRUE,
  lightThreshold = 10
)
}
\arguments{
\item{personfolder}{folder in which we expect all the pdk and Withings output folder.}

\item{desiredtz}{timezone (character) in Europe/London format}

\item{overwrite}{boolean if TRUE then it will reload all the data (default FALSE).}

\item{outputfolder}{the name of the RData file where all extracted data is stored.}

\item{ignore.light}{Boolean whether to ignore the light data}

\item{lightThreshold}{Threshold for light value above which light is stored}
}
\value{
outputfolder the name of the RData file where all extracted data is stored.
}
\description{
preprocess
}
