% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testplot.R
\name{testplot}
\alias{testplot}
\title{testplot}
\usage{
testplot(histfile, csvfile)
}
\arguments{
\item{histfile}{file where to store the plot}

\item{csvfile}{csvfile with the aggregated data}
}
\value{
No output, a png file is stored in the outputfolder
}
\description{
testplot
}
